+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## Unreleased
+
+* Added: WSMan TrustedHosts list manipulation commands
+
+
 ## 1.2.0 - 2017-07-20
 
 * Added: Invoke-PowerShell function with (alias: posh)